requirements.txt
asgiref==3.9.1
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.10.0
django-csp==4.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysql-connector-python==9.4.0
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
sqlparse==0.5.3
